Code for matrix inversion using cache
makeCacheMatrix creates a matrix like an object with properties that can
mantain his status. The variable 'x' is replaced by 'mtrx' because is a
more descriptive name.

cacheSolve obtains the inverse of the matrix, but if the inverse has
already been calculated simply uses the cache. 'x' has been replaced by
'mt' because is a more descriptive name.